Added option to use QUBO energy equation during output, added floating point error tolerance to physical validity checks

+++ global.h
@@ -16,7 +16,7 @@
 
 #include "logger.h"
 
-typedef double FPType;
+typedef float FPType;
 
 namespace global {
